% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outliers.R
\name{find_outliers}
\alias{find_outliers}
\title{Flag outliers in a numeric column of a 311 dataset}
\usage{
find_outliers(data, column = "minutes_to_close", k = 1.5)
}
\arguments{
\item{data}{311 data frame.}

\item{column}{Column name to check for outliers.}

\item{k}{Multiplier for the IQR. Default is 1.5.}
}
\value{
Logical vector, TRUE where x is considered an outlier.
}
\description{
Flag outliers in a numeric column of a 311 dataset
}
