% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_requests.R
\name{plot.civics311_summary}
\alias{plot.civics311_summary}
\title{Allow base plot() to work on civics311 summary objects.}
\usage{
\method{plot}{civics311_summary}(x, ...)
}
\arguments{
\item{x}{A civics311_summary.}

\item{...}{Additional arguments (unused).}
}
\description{
Allow base plot() to work on civics311 summary objects.
}
