% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_requests.R
\name{plot_requests}
\alias{plot_requests}
\title{Plot request counts from either a summary or raw data}
\usage{
plot_requests(x, ...)
}
\arguments{
\item{x}{A civics311_summary object or a raw 311 data frame.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A ggplot object showing request counts.
}
\description{
This function is a generic that dispatches to methods for
civics311_summary objects or raw data frames. If given raw data,
the function will summarize the data first.
}
