% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_requests.R
\name{print.civics311_summary}
\alias{print.civics311_summary}
\title{Print method for civics311_summary}
\usage{
\method{print}{civics311_summary}(x, ...)
}
\arguments{
\item{x}{A civics311_summary object.}

\item{...}{Additional arguments passed to print (unused).}
}
\description{
Print method for civics311_summary
}
