% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_requests.R
\name{summarize_requests}
\alias{summarize_requests}
\title{Summarize 311 requests by a grouping variable}
\usage{
summarize_requests(data, group_vars = "request_type")
}
\arguments{
\item{data}{A cleaned 311 data frame.}

\item{group_vars}{Character vector of columns to group by.}
}
\value{
A civics311_summary object.
}
\description{
Summarize 311 requests by a grouping variable
}
